api/requirements.txt
google-cloud-bigquery
google-auth
fastapi
uvicorn == 0.32.1
db_dtypes